+++ components/screenstreaming/impl/src/main/java/com/flipperdevices/screenstreaming/impl/composable/controls/ComposableFlipperDPad.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.screenstreaming.impl.composable
+package com.flipperdevices.screenstreaming.impl.composable.controls
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -25,6 +25,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.core.ui.theme.LocalPallet
+import com.flipperdevices.screenstreaming.impl.composable.ButtonEnum
 
 private const val BUTTON_WEIGHT = 0.3f